a3: add apple regen

+++ assignment3/lt-project/shared/src/main/scala/DSL/Shapes.scala
@@ -71,7 +71,7 @@ case class Grid(width: Int, height: Int, wall_top: Array[Wall],
 }
 
 
-case class Point(var x: Int, var y: Int){
+case class Point(var x: Double, var y: Double){
   def +(other: Point): Point = Point(x + other.x, y + other.y)
   def -(other: Point): Point = Point(x - other.x, y - other.y)
   def <(other: Point): Boolean = x < other.x && y < other.y
